Hotfix: Updated project ID's due to website bug
The website has a bug where a project cannot be deleted. To address this issue, I changed the Project ID's on Rolando's code to accurately reflect failed tests due to the bug. The code will still be unable to delete the projects, but it will be able to create different projects instead of running into "Project ID already exists" issue.